@@ -4,7 +4,7 @@ version: 0.1.5
 homepage: https://github.com/aloisdeniel/media_gallery
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-133.2.beta <3.0.0"
   flutter: ">=1.22.0 <2.0.0"
 
 dependencies:
