@@ -3,7 +3,7 @@ part of media_gallery;
 /// Fetches the given media thumbnail from the gallery.
 class MediaThumbnailProvider extends ImageProvider<MediaThumbnailProvider> {
   const MediaThumbnailProvider({
-    @required this.media,
+    required this.media,
   }) : assert(media != null);
 
   final Media media;
@@ -14,7 +14,7 @@ class MediaThumbnailProvider extends ImageProvider<MediaThumbnailProvider> {
       codec: _loadAsync(key, decode),
       scale: 1.0,
       informationCollector: () sync* {
-        yield ErrorDescription('Id: ${media?.id}');
+        yield ErrorDescription('Id: ${media.id}');
       },
     );
   }
@@ -23,9 +23,9 @@ class MediaThumbnailProvider extends ImageProvider<MediaThumbnailProvider> {
       MediaThumbnailProvider key, DecoderCallback decode) async {
     assert(key == this);
     final bytes = await media.getThumbnail();
-    if (bytes.length == 0) return null;
+    if (bytes.length == 0) return await decode([] as Uint8List);
 
-    return await decode(bytes);
+    return await decode(bytes as Uint8List);
   }
 
   @override
@@ -37,21 +37,21 @@ class MediaThumbnailProvider extends ImageProvider<MediaThumbnailProvider> {
   bool operator ==(dynamic other) {
     if (other.runtimeType != runtimeType) return false;
     final MediaThumbnailProvider typedOther = other;
-    return media?.id == typedOther.media?.id;
+    return media.id == typedOther.media.id;
   }
 
   @override
-  int get hashCode => media?.id?.hashCode ?? 0;
+  int get hashCode => media.id.hashCode;
 
   @override
-  String toString() => '$runtimeType("${media?.id}")';
+  String toString() => '$runtimeType("${media.id}")';
 }
 
 /// Fetches the given media collection thumbnail from the gallery.
 class MediaCollectionThumbnailProvider
     extends ImageProvider<MediaCollectionThumbnailProvider> {
   const MediaCollectionThumbnailProvider({
-    @required this.collection,
+    required this.collection,
   }) : assert(collection != null);
 
   final MediaCollection collection;
@@ -62,7 +62,7 @@ class MediaCollectionThumbnailProvider
       codec: _loadAsync(key, decode),
       scale: 1.0,
       informationCollector: () sync* {
-        yield ErrorDescription('Id: ${collection?.id}');
+        yield ErrorDescription('Id: ${collection.id}');
       },
     );
   }
@@ -71,9 +71,9 @@ class MediaCollectionThumbnailProvider
       MediaCollectionThumbnailProvider key, DecoderCallback decode) async {
     assert(key == this);
     final bytes = await collection.getThumbnail();
-    if (bytes == null || bytes.length == 0) return null;
+    if (bytes == null || bytes.length == 0) return await decode([] as Uint8List);
 
-    return await decode(bytes);
+    return await decode(bytes as Uint8List);
   }
 
   @override
@@ -86,14 +86,14 @@ class MediaCollectionThumbnailProvider
   bool operator ==(dynamic other) {
     if (other.runtimeType != runtimeType) return false;
     final MediaCollectionThumbnailProvider typedOther = other;
-    return collection?.id == typedOther.collection?.id;
+    return collection.id == typedOther.collection.id;
   }
 
   @override
-  int get hashCode => collection?.id?.hashCode ?? 0;
+  int get hashCode => collection.id.hashCode;
 
   @override
-  String toString() => '$runtimeType("${collection?.id}")';
+  String toString() => '$runtimeType("${collection.id}")';
 }
 
 /// Fetches the given media image thumbnail from the gallery.
@@ -101,7 +101,7 @@ class MediaCollectionThumbnailProvider
 /// The given [media] must be of media type [MediaType.image].
 class MediaImageProvider extends ImageProvider<MediaImageProvider> {
   MediaImageProvider({
-    @required this.media,
+    required this.media,
   })  : assert(media != null),
         assert(media.mediaType == MediaType.image);
 
@@ -113,7 +113,7 @@ class MediaImageProvider extends ImageProvider<MediaImageProvider> {
       codec: _loadAsync(key, decode),
       scale: 1.0,
       informationCollector: () sync* {
-        yield ErrorDescription('Id: ${media?.id}');
+        yield ErrorDescription('Id: ${media.id}');
       },
     );
   }
@@ -122,10 +122,10 @@ class MediaImageProvider extends ImageProvider<MediaImageProvider> {
       MediaImageProvider key, DecoderCallback decode) async {
     assert(key == this);
     final file = await media.getFile();
-    if (file == null) return null;
+    if (file == null) return await decode([] as Uint8List);
 
     final bytes = await file.readAsBytes();
-    if (bytes.lengthInBytes == 0) return null;
+    if (bytes.lengthInBytes == 0) return await decode([] as Uint8List);
 
     return await decode(bytes);
   }
@@ -139,12 +139,12 @@ class MediaImageProvider extends ImageProvider<MediaImageProvider> {
   bool operator ==(dynamic other) {
     if (other.runtimeType != runtimeType) return false;
     final MediaImageProvider typedOther = other;
-    return media?.id == typedOther.media?.id;
+    return media.id == typedOther.media.id;
   }
 
   @override
-  int get hashCode => media?.id?.hashCode ?? 0;
+  int get hashCode => media.id.hashCode;
 
   @override
-  String toString() => '$runtimeType("${media?.id}")';
+  String toString() => '$runtimeType("${media.id}")';
 }
